requirements.txt
pillow==10.4.0
just_playback==0.1.8
TCPLib~=2.0